#95 hotfix: plan Plan삭제 PlanData type으로 연결
