--- src/types/plan.ts
@@ -1,7 +1,0 @@
-// 밈 테스트에서 사용하는 Plan
-export interface Plan {
-  rank: number;
-  title: string;
-  subtitle: string;
-  detail: string;
-}
